Redirect to origin even if --skip-provider-button
The normal behavior is that if a user is going to page
https://hostname/x but is not authenticated, they are sent to login and
then redirected back to http://hostname/x.

I discovered that if the flag --skip-provider-button is given, then
after login the user ends up at http://hostname/. I believe this to be a
bug.

This change fixes that bug and adds a test.